chore(test): consolidate `LockMode.NONE` regression tests into `GH7702.test.ts`
